refactor(baro-rs): decompose home page into per-act subcomponents
The 636-line BaroRsHome becomes a 66-line orchestrator plus five
focused subcomponents:

  BaroHero.vue      — oscilloscope canvas + hero stack (owns rAF loop)
  BaroOrigin.vue    — origin copy + sticky spec sidebar
  BaroDrivers.vue   — driver list with per-driver --dc accent
  BaroArch.vue      — architecture SVG with computed edges
  BaroLog.vue       — commit log

Design thesis (page-baro-rs skill) is preserved as-is:
  - writing-mode vertical-rl gutter numbering via `.act::before`
  - instrument-dashboard palette (amber / copper / graphite) on root
  - grid background + asymmetric gutters
  - pulse keyframe for the hero status dot
  - SVG arch-edge connectors unchanged

The shared `.act` chrome rule (padding + border + writing-mode
gutter) moves to BaroRsHome.vue with `:deep()` so it still applies
to child-rendered section elements. Same pattern for the `.section-label`
helper.

Oscilloscope rAF moves into BaroHero where the canvas lives.
`useLenis()` stays at the orchestrator level where scroll is owned.

No visual change: 'bun run test:visual -- --grep baro-rs' green on
all three viewports for both home and driver routes.
